refactor(friendStock): style 새파일로 분리
