+++ components/community/Friends/style.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const CommunitySection = styled.section`
+  display: flex;
+  gap: 7.1rem;
+`;
+
+export { CommunitySection };
